+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add support for `diffie-hellman-group1-sha1` key exchange under `insecure-crypto` feature.
+- Add support for `3des-cbc` cipher under `insecure-crypto` feature.
 
 ## 0.2.3 (2024-07-22)